userbot: Fix session invalidation issues on Heroku
Commit [1] introduced a workaround for the issue, but it didn't really
fix the issue and was just a dirty hack.

Every time Heroku migrates the dyno, it issues SIGTERM to all processes.
By handling it gracefully and disconnecting the Telegram session, the
bot can be shut down properly. The session won't be accessed from two
instances at once.

I've still kept the timeout function from [1] in case we do need it down
the line, but the bot has been working fine in my tests without it.

[1]: https://github.com/KenHV/KensurBot/commit/824ed7044b6519ad456e6dfe15bedeb0e7d76d77

Change-Id: Ie3425076a8f03f1f9701126b3d59da896d41312e
Signed-off-by: KenHV <yo@kenharris.xyz>